jee_math_solver/requirements.txt
google-generativeai>=0.3.0
sympy>=1.12
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
